added Email conditional rendering based on role
