+++ src/components/Email/EmailAdminView.jsx
@@ -1,0 +1,11 @@
+import React, { useState, useEffect } from 'react';
+
+function EmailAdminView() {
+  return (
+    <div>
+      <h1>Email Admin</h1>
+    </div>
+  );
+}
+
+export default EmailAdminView;
